refactor: extract as useFakeSignOut custom hook
#70
