+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useLogin';
 export * from './useInput';
 export * from './useCards';
+export * from './useFakeSignOut';
